fix code after feedback
