+++ src/app/components/tables/subject-table/subject-table.component.html
@@ -1,29 +1,28 @@
 <div class="table-container mat-elevation-z8">
-
-  <md-table #table [dataSource]="dataSource">
+  <md-table #table [dataSource]="dataSource" [trackBy]="trackById">
 
     <!-- Patient ID Column -->
     <ng-container cdkColumnDef="subjectId">
       <md-header-cell *cdkHeaderCellDef> Patient ID </md-header-cell>
-      <md-cell *cdkCellDef="let row"> {{row.subjectId}} </md-cell>
+      <md-cell *cdkCellDef="let row"> {{ row.subjectId }} </md-cell>
     </ng-container>
 
     <!-- Active Column -->
     <ng-container cdkColumnDef="active">
       <md-header-cell *cdkHeaderCellDef> Active </md-header-cell>
-      <md-cell *cdkCellDef="let row"> {{row.active}} </md-cell>
+      <md-cell *cdkCellDef="let row"> {{ row.active }} </md-cell>
     </ng-container>
 
     <!-- Effective Time Frame Start Column -->
     <ng-container cdkColumnDef="startdate">
       <md-header-cell *cdkHeaderCellDef> Start Date Time </md-header-cell>
-      <md-cell *cdkCellDef="let row"> {{row.effectiveTimeFrame.endDateTime}} </md-cell>
+      <md-cell *cdkCellDef="let row"> {{ row.effectiveTimeFrame.endDateTime | date: 'medium' }} </md-cell>
     </ng-container>
 
     <!-- Effective Time Frame End Column -->
     <ng-container cdkColumnDef="enddate">
       <md-header-cell *cdkHeaderCellDef> End Date Time </md-header-cell>
-      <md-cell *cdkCellDef="let row"> {{row.effectiveTimeFrame.startDateTime}} </md-cell>
+      <md-cell *cdkCellDef="let row"> {{ row.effectiveTimeFrame.startDateTime | date: 'medium' }} </md-cell>
     </ng-container>
 
     <!-- Sources Column -->
@@ -39,7 +38,7 @@
   </md-table>
 
   <md-paginator #paginator
-                [length]="dataSource.mydata.length"
+                [length]="dataSource.data.length"
                 [pageIndex]="0"
                 [pageSize]="10"
                 [pageSizeOptions]="[5, 10, 25, 100]">
