+++ src/app/components/tables/subject-table/subject-table.component.scss
@@ -1,13 +1,11 @@
 @import '../../../../styles/common';
 
 .table-container {
-  float: left;
   clear: none;
   overflow: hidden;
 }
 
 .table-container md-table {
-  float: left;
   clear: none;
   width: 100%;
 }
